+++ UnityScripts/scripts/Objects/Food.cs
@@ -203,7 +203,7 @@ public class Food : object_base {
 		}	
 	}
 
-	public override bool ApplyAttack (int damage)
+	public override bool ApplyAttack (short damage)
 	{
 		objInt().quality-=damage;
 		if (objInt().quality<=0)
